Fixed status bar
The issue was the uptime output changed because the number of
columns change over time.

The solution was to display the last column only no matter how
many columns there were using $NF and field seperator ':'. The
dollar sign in $NF had to be escaped with a backslash to function
in the tmux status bar.